Correcion de models, creacion de controladores, creacion de item e nav

+++ TrabajoPractico_Integrador/Models/Detalle_venta.cs
@@ -5,8 +5,8 @@ namespace TrabajoPractico_Integrador.Models
     {
         public int Id { get; set; }
         public int Cantidad { get; set; }
-        public int PrecioUnitario { get; set; }
-        public int Subtotal { get; set; }
+        public decimal PrecioUnitario { get; set; }
+        public decimal Subtotal { get; set; }
         public int ProductoID { get; set; }
         public int VentaID { get; set; }
         public Producto Producto { get; set; }
