+++ TrabajoPractico_Integrador/Models/Producto.cs
@@ -8,7 +8,7 @@ namespace TrabajoPractico_Integrador.Models
         public int Id { get; set; }
         public string Descripcion { get; set; }
         public int Stock { get; set; }
-        public int PrecioVenta { get; set; }
+        public decimal PrecioVenta { get; set; }
         public ICollection <Detalle_venta> Detalles { get; set; }
         public Producto() {
         }
